Fix linters

+++ app/controllers/comments_controller.rb
@@ -6,20 +6,17 @@ class CommentsController < ApplicationController
     end
   end
 
-	def destroy
-		comment = Comment.includes(post: [:user]).find(params[:id])
-		post = comment.post
-		if can? :destroy, comment
-			comment.destroy
-		end
-		flash[:error] = 'cannot delete comment'
-		redirect_to "/users/#{post.user.id}/posts/#{post.id}/"
-	end
-
+  def destroy
+    comment = Comment.includes(post: [:user]).find(params[:id])
+    post = comment.post
+    comment.destroy if can? :destroy, comment
+    flash[:error] = 'cannot delete comment'
+    redirect_to "/users/#{post.user.id}/posts/#{post.id}/"
+  end
 
   def create
     comment = Comment.new(params.require(:comment).permit(:text))
-		post = Post.find(params[:post_id])
+    post = Post.find(params[:post_id])
     comment.user = current_user
     comment.post = post
 
@@ -30,7 +27,7 @@ class CommentsController < ApplicationController
         else
           flash[:error] = 'Comment faild to create'
         end
-				redirect_to "/users/#{post.user.id}/posts/#{post.id}/"
+        redirect_to "/users/#{post.user.id}/posts/#{post.id}/"
       end
     end
   end
